+++ testdata/link/dst/main.c
@@ -3,6 +3,6 @@
 // -- Implementation --
 
 int main(void) {
-    (void)math_Sqrt(2);
+    (void)math_Sqrt(2.0);
     return 0;
 }