+++ testdata/packages/dst/main.c
@@ -3,7 +3,7 @@
 // -- Implementation --
 
 int main(void) {
-    double a = geom_RectArea(5, 10);
+    double a = geom_RectArea(5.0, 10.0);
     (void)a;
     (void)geom_Pi;
     return 0;